Add an option to use the American RBDS standard.
RBDS is based on the European RDS standard, with just a couple of incompatible differences to keep life interesting.

See http://www.nrscstandards.org/sg/nrsc-4-b.pdf for the standard.  Unfortunately it incorporates the RDS standard by reference, and it looks like there's a paywall on that, so I'll take the existing RDS code on faith :)